+++ previous-versions/banking/src/codatbanking/models/operations/get_transaction_category.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import transactioncategory as shared_transactioncategory
+from ...models.shared import transactioncategory as shared_transactioncategory
 from typing import Optional
 
 
@@ -26,8 +25,6 @@ class GetTransactionCategoryResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     transaction_category: Optional[shared_transactioncategory.TransactionCategory] = dataclasses.field(default=None)